requirements.txt
Django==2.2.1
pytz==2019.1
sqlparse==0.3.0
redis==3.2.1
django-redis==4.10.0
python-decouple==3.1
